server/requirements.txt
flask
flask-cors
gunicorn
mysql-connector-python
python-dotenv
PyPDF2
python-docx
transformers
torch
psutil
Pillow>=10.0.0